Upgrade to Halogen 0.5.0, remove Breadcrumps & Navbar; fix #7
This commit upgrades this library to Halogen 0.5.0, and removes a couple
of wrappers that are no longer compatible with Halogen following
discussion with @garyb; these may be rewritten in the future as
components, but it makes sense for this library to remain a "barebones"
typed interface to the Bootstrap classes, rather than a framework in its
own right.

I have also added a self-contained build (package.json, .travis.yml).